Stage10 4,6,7,8번 추가

+++ BAEKJOON_Problem/Stage10/Stage10_04.cs
@@ -4,17 +4,14 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BAEKJOON_Problem.Stage10
-{
-    internal class Stage10_04
-    {
+namespace BAEKJOON_Problem.Stage10 {
+    internal class Stage10_04 {
 
-        static void Main(string[] args)
-        {
+        static void Main(string[] args) {
 
-            int n = int.Parse(Console.ReadLine());
+            long n = int.Parse(Console.ReadLine());
 
-            int result = n * 4;
+            long result = n * 4;
 
             Console.WriteLine(result);
